+++ src/clmf_ref.cpp
@@ -1,8 +1,8 @@
 #include "clmf_ref.h"
 
-void choldc1(int n, VALUE_TYPE** a, VALUE_TYPE* p) {
-    for (int i = 0; i < n; ++i) {
-        for (int j = i; j < n; ++j) {
+void choldc1(unsigned n, VALUE_TYPE** a, VALUE_TYPE* p) {
+    for (unsigned i = 0; i < n; ++i) {
+        for (unsigned j = i; j < n; ++j) {
             VALUE_TYPE sum = a[i][j];
             for (int k = i - 1; k >= 0; --k) {
                 sum -= a[i][k] * a[j][k];
@@ -19,15 +19,15 @@ void choldc1(int n, VALUE_TYPE** a, VALUE_TYPE* p) {
     }
 }
 
-void choldcsl(int n, VALUE_TYPE** A) {
+void choldcsl(unsigned n, VALUE_TYPE** A) {
     VALUE_TYPE* p = (VALUE_TYPE*) malloc(n * sizeof(VALUE_TYPE));
     choldc1(n, A, p);
 
-    for (int i = 0; i < n; ++i) {
+    for (unsigned i = 0; i < n; ++i) {
         A[i][i] = 1 / p[i];
-        for (int j = i + 1; j < n; ++j) {
+        for (unsigned j = i + 1; j < n; ++j) {
             VALUE_TYPE sum = 0;
-            for (int k = i; k < j; ++k) {
+            for (unsigned k = i; k < j; ++k) {
                 sum -= A[j][k] * A[k][i];
             }
             A[j][i] = sum / p[j];
@@ -36,37 +36,37 @@ void choldcsl(int n, VALUE_TYPE** A) {
     free(p);
 }
 
-void inverseMatrix_CholeskyMethod(int n, VALUE_TYPE** A) {
+void inverseMatrix_CholeskyMethod(unsigned n, VALUE_TYPE** A) {
     choldcsl(n, A);
-    for (int i = 0; i < n; ++i) {
-        for (int j = i + 1; j < n; ++j) {
+    for (unsigned i = 0; i < n; ++i) {
+        for (unsigned j = i + 1; j < n; ++j) {
             A[i][j] = 0.0;
         }
     }
-    for (int i = 0; i < n; i++) {
+    for (unsigned i = 0; i < n; i++) {
         A[i][i] *= A[i][i];
-        for (int k = i + 1; k < n; ++k) {
+        for (unsigned k = i + 1; k < n; ++k) {
             A[i][i] += A[k][i] * A[k][i];
         }
-        for (int j = i + 1; j < n; ++j) {
-            for (int k = j; k < n; ++k) {
+        for (unsigned j = i + 1; j < n; ++j) {
+            for (unsigned k = j; k < n; ++k) {
                 A[i][j] += A[k][i] * A[k][j];
             }
         }
     }
-    for (int i = 0; i < n; ++i) {
-        for (int j = 0; j < i; ++j) {
+    for (unsigned i = 0; i < n; ++i) {
+        for (unsigned j = 0; j < i; ++j) {
             A[i][j] = A[j][i];
         }
     }
 }
 
-void Mt_byM_multiply(int i, int j, VALUE_TYPE** M, VALUE_TYPE** Result) {
+void Mt_byM_multiply(unsigned i, unsigned j, VALUE_TYPE** M, VALUE_TYPE** Result) {
     VALUE_TYPE SUM;
-    for (int I = 0; I < j; ++I) {
-        for (int J = I; J < j; ++J) {
+    for (unsigned I = 0; I < j; ++I) {
+        for (unsigned J = I; J < j; ++J) {
             SUM = 0.0;
-            for (int K = 0; K < i; ++K) {
+            for (unsigned K = 0; K < i; ++K) {
                 //printf("%.3f %.3f\n", M[K][I], M[K][J]);
                 SUM += M[K][I] * M[K][J];
             }
@@ -79,7 +79,7 @@ void Mt_byM_multiply(int i, int j, VALUE_TYPE** M, VALUE_TYPE** Result) {
 #define kind dynamic,500
 
 void clmf_ref(SparseMatrix& R, MatData& W, MatData& H, TestData& T,parameter& param) {
-    int k = (int) param.k;
+    unsigned k = param.k;
 
     int num_threads_old = omp_get_num_threads();
     omp_set_num_threads(param.threads);
@@ -100,7 +100,7 @@ void clmf_ref(SparseMatrix& R, MatData& W, MatData& H, TestData& T,parameter& pa
             if (omegaSize > 0) {
                 VALUE_TYPE* subVector = (VALUE_TYPE*) malloc(k * sizeof(VALUE_TYPE));
                 subMatrix = (VALUE_TYPE**) malloc(k * sizeof(VALUE_TYPE*));
-                for (int i = 0; i < k; ++i) {
+                for (unsigned i = 0; i < k; ++i) {
                     subMatrix[i] = (VALUE_TYPE*) malloc(k * sizeof(VALUE_TYPE));
                 }
 
@@ -113,7 +113,7 @@ void clmf_ref(SparseMatrix& R, MatData& W, MatData& H, TestData& T,parameter& pa
                 Mt_byM_multiply(omegaSize, k, H_Omega, subMatrix);
 
                 //add lambda to diag of sub-matrix
-                for (int c = 0; c < k; c++) {
+                for (unsigned c = 0; c < k; c++) {
                     subMatrix[c][c] = subMatrix[c][c] + param.lambda;
                 }
 
@@ -122,7 +122,7 @@ void clmf_ref(SparseMatrix& R, MatData& W, MatData& H, TestData& T,parameter& pa
 
 
                 //sparse multiplication
-                for (int c = 0; c < k; ++c) {
+                for (unsigned c = 0; c < k; ++c) {
                     subVector[c] = 0;
                     for (unsigned idx = R.get_csr_row_ptr()[Rw]; idx < R.get_csr_row_ptr()[Rw + 1]; ++idx) {
                         subVector[c] += R.get_csr_val()[idx] * H[R.get_csr_col_indx()[idx]][c];
@@ -130,22 +130,22 @@ void clmf_ref(SparseMatrix& R, MatData& W, MatData& H, TestData& T,parameter& pa
                 }
 
                 //multiply subVector by subMatrix
-                for (int c = 0; c < k; ++c) {
+                for (unsigned c = 0; c < k; ++c) {
                     Wr[c] = 0;
-                    for (int subVid = 0; subVid < k; ++subVid) {
+                    for (unsigned subVid = 0; subVid < k; ++subVid) {
                         Wr[c] += subVector[subVid] * subMatrix[c][subVid];
                     }
                 }
 
 
-                for (int i = 0; i < k; ++i) {
+                for (unsigned i = 0; i < k; ++i) {
                     free(subMatrix[i]);
                 }
                 free(subMatrix);
                 free(subVector);
                 free(H_Omega);
             } else {
-                for (int c = 0; c < k; ++c) {
+                for (unsigned c = 0; c < k; ++c) {
                     Wr[c] = 0.0;
                     //printf("%.3f ", Wr[c]);
                 }
@@ -163,7 +163,7 @@ void clmf_ref(SparseMatrix& R, MatData& W, MatData& H, TestData& T,parameter& pa
             if (omegaSize > 0) {
                 VALUE_TYPE* subVector = (VALUE_TYPE*) malloc(k * sizeof(VALUE_TYPE));
                 subMatrix = (VALUE_TYPE**) malloc(k * sizeof(VALUE_TYPE*));
-                for (int i = 0; i < k; ++i) {
+                for (unsigned i = 0; i < k; ++i) {
                     subMatrix[i] = (VALUE_TYPE*) malloc(k * sizeof(VALUE_TYPE));
                 }
 
@@ -176,7 +176,7 @@ void clmf_ref(SparseMatrix& R, MatData& W, MatData& H, TestData& T,parameter& pa
                 Mt_byM_multiply(omegaSize, k, W_Omega, subMatrix);
 
                 //add lambda to diag of sub-matrix
-                for (int c = 0; c < k; c++) {
+                for (unsigned c = 0; c < k; c++) {
                     subMatrix[c][c] = subMatrix[c][c] + param.lambda;
                 }
 
@@ -185,7 +185,7 @@ void clmf_ref(SparseMatrix& R, MatData& W, MatData& H, TestData& T,parameter& pa
 
 
                 //sparse multiplication
-                for (int c = 0; c < k; ++c) {
+                for (unsigned c = 0; c < k; ++c) {
                     subVector[c] = 0;
                     for (unsigned idx = R.get_csc_col_ptr()[Rh]; idx < R.get_csc_col_ptr()[Rh + 1]; ++idx) {
                         subVector[c] += R.get_csc_val()[idx] * W[R.get_csc_row_indx()[idx]][c];
@@ -193,22 +193,22 @@ void clmf_ref(SparseMatrix& R, MatData& W, MatData& H, TestData& T,parameter& pa
                 }
 
                 //multiply subVector by subMatrix
-                for (int c = 0; c < k; ++c) {
+                for (unsigned c = 0; c < k; ++c) {
                     Hr[c] = 0;
-                    for (int subVid = 0; subVid < k; ++subVid) {
+                    for (unsigned subVid = 0; subVid < k; ++subVid) {
                         Hr[c] += subVector[subVid] * subMatrix[c][subVid];
                     }
                 }
 
 
-                for (int i = 0; i < k; ++i) {
+                for (unsigned i = 0; i < k; ++i) {
                     free(subMatrix[i]);
                 }
                 free(subMatrix);
                 free(subVector);
                 free(W_Omega);
             } else {
-                for (int c = 0; c < k; ++c) {
+                for (unsigned c = 0; c < k; ++c) {
                     Hr[c] = 0.0;
                 }
             }
